WOrking attack logic

+++ include/User.hpp
@@ -39,7 +39,7 @@ class User{
 
 
         //void fight_user(User other);
-        void deploy(int index, sf::Vector2i d_loc); //abstract class
+        void deploy(int index, sf::Vector2f d_loc); //abstract class
 
 };
